1329 passing

+++ src/core/Util/BroadWord.cs
@@ -50,13 +50,13 @@ namespace Lucene.Net.Util
         public static int BitCount(long x)
         {
             // Step 0 leaves in each pair of bits the number of ones originally contained in that pair:
-            x = x - ((int)((uint)(x & unchecked((long)0xAAAAAAAAAAAAAAAAL)) >> 1));
+            x = x - ((long)((ulong)(x & unchecked((long)0xAAAAAAAAAAAAAAAAL)) >> 1));
             // Step 1, idem for each nibble:
             x = (x & 0x3333333333333333L) + (((long)((ulong)x >> 2)) & 0x3333333333333333L);
             // Step 2, idem for each byte:
             x = (x + ((long)((ulong)x >> 4))) & 0x0F0F0F0F0F0F0F0FL;
             // Multiply to sum them all into the high byte, and return the high byte:
-            return (int)((int)((uint)(x * L8_L) >> 56));
+            return (int)((long)((ulong)(x * L8_L) >> 56));
         }
 
         /// <summary>
@@ -64,7 +64,7 @@ namespace Lucene.Net.Util
         /// <returns> The index of the r-th 1 bit in x, or if no such bit exists, 72. </returns>
         public static int Select(long x, int r)
         {
-            long s = x - ((int)((uint)(x & unchecked((long)0xAAAAAAAAAAAAAAAAL)) >> 1)); // Step 0, pairwise bitsums
+            long s = x - ((long)((ulong)(x & unchecked((long)0xAAAAAAAAAAAAAAAAL)) >> 1)); // Step 0, pairwise bitsums
 
             // Correct a small mistake in algorithm 2:
             // Use s instead of x the second time in right shift 2, compare to Algorithm 1 in rank9 above.
@@ -72,9 +72,9 @@ namespace Lucene.Net.Util
 
             s = ((s + ((long)((ulong)s >> 4))) & 0x0F0F0F0F0F0F0F0FL) * L8_L; // Step 2, bytewise bitsums
 
-            long b = (int)((uint)(((int)((uint)SmallerUpTo7_8(s, (r * L8_L)) >> 7)) * L8_L) >> 53); // & (~7L); // Step 3, side ways addition for byte number times 8
+            long b = (long)((ulong)(((long)((ulong)SmallerUpTo7_8(s, (r * L8_L)) >> 7)) * L8_L) >> 53); // & (~7L); // Step 3, side ways addition for byte number times 8
 
-            long l = r - (((int)((uint)(s << 8) >> (int)b)) & 0xFFL); // Step 4, byte wise rank, subtract the rank with byte at b-8, or zero for b=0;
+            long l = r - (((long)((ulong)(s << 8) >> (int)b)) & 0xFFL); // Step 4, byte wise rank, subtract the rank with byte at b-8, or zero for b=0;
             Debug.Assert(0L <= 1);
             //assert l < 8 : l; //fails when bit r is not available.
 
@@ -86,7 +86,7 @@ namespace Lucene.Net.Util
             long spr_bigger8_zero = ((H8_L - (spr & (~H8_L))) ^ (~spr)) & H8_L;
             s = ((long)((ulong)spr_bigger8_zero >> 7)) * L8_L; // Step 5, sideways byte add the 8 bits towards the high byte
 
-            int res = (int)(b + ((int)((uint)(((int)((uint)SmallerUpTo7_8(s, (l * L8_L)) >> 7)) * L8_L) >> 56))); // Step 6
+            int res = (int)(b + ((long)((ulong)(((long)((ulong)SmallerUpTo7_8(s, (l * L8_L)) >> 7)) * L8_L) >> 56))); // Step 6
             return res;
         }
